Move RawWrite into crucible_protocol as well
